readd deleted functions by mistake

+++ dtrace-alloc-postprocess/main.c
@@ -1,4 +1,3 @@
-#include <_types/_uint64_t.h>
 #include <errno.h>
 #include <fcntl.h>
 #include <stdlib.h>
@@ -505,7 +504,7 @@ int main(int argc, char* argv[]) {
       pg_log_fatal(&logger, errno, "Failed to open file %s: %s", argv[1],
                    strerror(errno));
     }
-    int64_t ret = 0;
+    int ret = 0;
     if ((ret = pg_read_file(pg_heap_allocator(), argv[1], &file_data)) != 0) {
       pg_log_fatal(&logger, ret, "Failed to read file %s: %s", argv[1],
                    strerror(ret));
